LeafSet and minor cleanup

+++ include/mw/core/util/BitUtil.h
@@ -30,7 +30,7 @@ public:
         uint8_t count = 0;
         while (n)
         {
-            count += n & 1;
+            count += (uint8_t)(n & 1);
             n >>= 1;
         }
